Refactor: add json-rpc endpoint folder structure
