+++ packages/sources/json-rpc/src/adapter.ts
@@ -1,49 +1,15 @@
-import { Requester, Validator } from '@chainlink/ea-bootstrap'
-import { Config, ExecuteFactory, ExecuteWithConfig } from '@chainlink/types'
-import { DEFAULT_BASE_URL, ExtendedConfig, makeConfig } from './config'
+import { Builder } from '@chainlink/ea-bootstrap'
+import { ExecuteWithConfig, ExecuteFactory, AdapterRequest, APIEndpoint } from '@chainlink/types'
+import { makeConfig, ExtendedConfig } from './config'
+import * as endpoints from './endpoint'
 
-const inputParams = {
-  url: false,
-  method: false,
-  params: false,
-  requestId: false,
+export const execute: ExecuteWithConfig<ExtendedConfig> = async (request, context, config) => {
+  return Builder.buildSelector(request, context, config, endpoints)
 }
 
-// Export function to integrate with Chainlink node
-export const execute: ExecuteWithConfig<ExtendedConfig> = async (request, _, config) => {
-  const validator = new Validator(request, inputParams)
+export const endpointSelector = (request: AdapterRequest): APIEndpoint =>
+  Builder.selectEndpoint(request, makeConfig(), endpoints)
 
-  const jobRunID = validator.validated.id
-  const url = config.RPC_URL || validator.validated.data.url || DEFAULT_BASE_URL
-  const method = validator.validated.data.method || ''
-  const params = validator.validated.data.params
-  const requestId = validator.validated.data.requestId || jobRunID
-
-  const data = {
-    id: requestId,
-    jsonrpc: '2.0',
-    method,
-    params,
-  }
-
-  const options = {
-    ...config.api,
-    url,
-    method: 'POST',
-    headers: {
-      ...config.api.headers,
-      'Content-Type': 'application/json',
-    },
-    // Remove undefined values
-    data: JSON.parse(JSON.stringify(data)),
-  }
-
-  const response = await Requester.request(options)
-  if (response.status >= 400) throw response.data.error
-
-  return Requester.success(request.id, response)
-}
-
-export const makeExecute: ExecuteFactory<Config> = (config) => {
+export const makeExecute: ExecuteFactory<ExtendedConfig> = (config) => {
   return async (request, context) => execute(request, context, config || makeConfig())
 }
